backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic
pydantic-settings
python-jose
python-multipart
httpx
playwright
apscheduler
google-generativeai
langchain
langchain-google-genai
langchain-mongodb
langchain-community
langgraph
pymupdf
requests
cloudinary
resend
twilio
Pillow
websockets  
beautifulsoup4  
lxml  
pymongo<4.11.0 
